requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
aiosqlite==0.16.1
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
bcrypt==3.2.0
cffi==1.14.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
fastapi-cache2==0.1.3.4
h11==0.12.0
hiredis==1.1.0
httptools==0.1.1
idna==3.1
iso8601==0.1.14
multidict==5.1.0
passlib==1.7.4
pycparser==2.20
pydantic==1.8.1
PyJWT==2.0.1
PyPika==0.44.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.4.1
six==1.15.0
starlette==0.13.6
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
yarl==1.6.3
